Planner overhaul with podset-planner-scheduler and controller. The planner also runs through some built-in predicates for pods that have been assigned nodes internally to decide on the eligible set while planning the pods belonging to a podset.
The scheduler plugin invokes the planner service during preScore
plugin invocation once default scheduler has run through all filters.

It then saves the plan based on whether a scheduling trigger was specified
and active or not.

The controller for the scheduling trigger is a quiet timer that changes
the trigger state to schedule after a period thus enabling the pods
in the podset to be planned.